Split CurveEditor out to TS file
